requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
ipython==8.18.1
google-generativeai==0.3.2
